+++ src/main/java/ua/holovchenko/module4web/database/repo/interfaces/RaceDAO.java
@@ -4,5 +4,5 @@ import ua.holovchenko.module4web.database.models.Race;
 
 public interface RaceDAO extends GenericDAO<Race> {
     Race getRaceInfo(String id);
-    int getRacesCount();
+    long getRacesCount();
 }
